refactor extractors and remove debuginfo warning
