+++ src/gt4py/next/program_processors/runners/dace_fieldview/program.py
@@ -15,10 +15,10 @@ from typing import Any, ClassVar, Optional, Sequence
 import dace
 import numpy as np
 
-from gt4py import eve
 from gt4py.next import backend as next_backend, common
 from gt4py.next.ffront import decorator
 from gt4py.next.iterator import ir as itir
+from gt4py.next.iterator.transforms import extractors as extractors
 from gt4py.next.otf import arguments, recipes, toolchain
 from gt4py.next.program_processors.runners.dace_common import utility as dace_utils
 from gt4py.next.type_system import type_specifications as ts
@@ -96,10 +96,14 @@ class Program(decorator.Program, dace.frontend.python.common.SDFGConvertible):
                 if len(horizontal_dims) == 1:
                     yield str(field.id), horizontal_dims[0]
 
-        input_fields = (field_params[name] for name in InputNamesExtractor.only_fields(program))
+        input_fields = (
+            field_params[name] for name in extractors.InputNamesExtractor.only_fields(program)
+        )
         sdfg.gt4py_program_input_fields = dict(single_horizontal_dim_per_field(input_fields))
 
-        output_fields = (field_params[name] for name in OutputNamesExtractor.only_fields(program))
+        output_fields = (
+            field_params[name] for name in extractors.OutputNamesExtractor.only_fields(program)
+        )
         sdfg.gt4py_program_output_fields = dict(single_horizontal_dim_per_field(output_fields))
 
         # TODO (ricoh): bring back sdfg.offset_providers_per_input_field.
@@ -189,79 +193,6 @@ class Program(decorator.Program, dace.frontend.python.common.SDFGConvertible):
 
     def __sdfg_signature__(self) -> tuple[Sequence[str], Sequence[str]]:
         return [p.id for p in self.past_stage.past_node.params], []
-
-
-class SymbolNameSetExtractor(eve.NodeVisitor):
-    """Extract a set of symbol names"""
-
-    def generic_visitor(self, node: itir.Node) -> set[str]:
-        input_fields: set[str] = set()
-        for child in eve.trees.iter_children_values(node):
-            input_fields |= self.visit(child)
-        return input_fields
-
-    @classmethod
-    def only_fields(cls, program: itir.Program) -> set[str]:
-        field_param_names = [
-            str(param.id) for param in program.params if isinstance(param.type, ts.FieldType)
-        ]
-        return {name for name in cls().visit(program) if name in field_param_names}
-
-
-class InputNamesExtractor(SymbolNameSetExtractor):
-    """Extract the set of symbol names passed into field operators within a program."""
-
-    def visit_Program(self, node: itir.Program) -> set[str]:
-        input_fields = set()
-        for stmt in node.body:
-            input_fields |= self.visit(stmt)
-        return input_fields
-
-    def visit_IfStmt(self, node: itir.IfStmt) -> set[str]:
-        input_fields = set()
-        for stmt in node.true_branch + node.false_branch:
-            input_fields |= self.visit(stmt)
-        return input_fields
-
-    def visit_Temporary(self, node: itir.Temporary) -> set[str]:
-        return set()
-
-    def visit_SetAt(self, node: itir.SetAt) -> set[str]:
-        return self.visit(node.expr)
-
-    def visit_FunCall(self, node: itir.FunCall) -> set[str]:
-        input_fields = set()
-        for arg in node.args:
-            input_fields |= self.visit(arg)
-        return input_fields
-
-    def visit_SymRef(self, node: itir.SymRef) -> set[str]:
-        return {str(node.id)}
-
-
-class OutputNamesExtractor(SymbolNameSetExtractor):
-    """Extract the set of symbol names written to within a program"""
-
-    def visit_Program(self, node: itir.Program) -> set[str]:
-        output_fields = set()
-        for stmt in node.body:
-            output_fields |= self.visit(stmt)
-        return output_fields
-
-    def visit_IfStmt(self, node: itir.IfStmt) -> set[str]:
-        output_fields = set()
-        for stmt in node.true_branch + node.false_branch:
-            output_fields |= self.visit(stmt)
-        return output_fields
-
-    def visit_Temporary(self, node: itir.Temporary) -> set[str]:
-        return set()
-
-    def visit_SetAt(self, node: itir.SetAt) -> set[str]:
-        return self.visit(node.target)
-
-    def visit_SymRef(self, node: itir.SymRef) -> set[str]:
-        return {str(node.id)}
 
 
 def _crosscheck_dace_parsing(dace_parsed_args: list[Any], gt4py_program_args: list[Any]) -> None:
